requirements.txt
Django~=4.2.7
djangorestframework~=3.14.0
python-decouple~=3.8.0
user-agents~=2.2.0
django-cors-headers~=3.14.0
Pillow~=10.1.0